rename module type to types
