+++ tests/unit/acquisition/test_function.py
@@ -71,7 +71,7 @@ from trieste.data import Dataset
 from trieste.models import ProbabilisticModel
 from trieste.objectives.single_objectives import BRANIN_MINIMUM, branin
 from trieste.space import Box
-from trieste.type import TensorType
+from trieste.types import TensorType
 from trieste.utils import DEFAULTS
 from trieste.utils.pareto import Pareto, get_reference_point
