+++ tests/util/model.py
@@ -22,7 +22,7 @@ import tensorflow_probability as tfp
 from tests.util.misc import SequenceN, quadratic
 from trieste.data import Dataset
 from trieste.models import ProbabilisticModel, TrainableProbabilisticModel
-from trieste.type import TensorType
+from trieste.types import TensorType
 
 
 def rbf() -> tfp.math.psd_kernels.ExponentiatedQuadratic:
