+++ trieste/types.py
